ENH : Major restructuring of code
