--- SciStreams/__init__.py
@@ -1,4 +1,0 @@
-# this reads from .config and gets config info automatically
-# from .interfaces.databroker.databases import databases
-# cadb = databases['cms:analysis']
-# cddb = databases['cms:analysis']
