--- SciStreams/analyses/XSAnalysis/__init__.py
@@ -1,16 +1,0 @@
-# -*- coding: utf-8 -*-
-###################################################################
-# SciAnalysis
-# version 0.5
-# Oct 2015
-###################################################################
-# Author: Kevin G. Yager
-# Affiliation: Brookhaven National Lab, Center for Functional Nanomaterials
-###################################################################
-# Description:
-#  SciAnalysis is a collection of Python tools for analyzing scientific
-# datasets.
-###################################################################
-#  This file is simply a placeholder to tell Python that this
-# directory should be parsed as a package.
-###################################################################
