+++ SciStreams/core/StreamDoc.py
