+++ SciStreams/core/streams.py
