+++ SciStreams/data/detectors.py
