+++ SciStreams/data/tests/test_Obstructions.py
@@ -1,4 +1,4 @@
-from SciStreams.analyses.XSAnalysis.Data import Obstruction
+from SciStreams.data.Obstructions import Obstruction
 import numpy as np
 from numpy.testing import assert_array_almost_equal
 
@@ -13,7 +13,7 @@ def test_obstruction():
     assert np.all(obs.mask[4:5, 6:9]==0)
 
     obs2 = Obstruction(mask, origin)
-    obs2 = obs2.rotate(180, rotation_offset=(10,10))
+    obs2.rotate(180, rotation_offset=(10,10))
     # the coordinates should shift this way, tested
     assert_array_almost_equal(obs2.origin, [-16, -17])
