+++ SciStreams/processing/partitioning.py
@@ -1,0 +1,29 @@
+import numpy as np
+
+def center2edge(centers, positive=True):
+    ''' Transform a set of bin centers to edges
+        This is useful for non-uniform bins.
+
+        Note : for the edges, an assumption is made. They are extended to half
+        the distance between the first two and last two points etc.
+
+        positive : make sure the edges are monotonically increasing
+    '''
+    midpoints = (centers[:-1] + centers[1:])*.5
+    dedge_left = centers[1]-centers[0]
+    dedge_right = centers[-1]-centers[-2]
+    left_edge = (centers[0] - dedge_left/2.).reshape(1)
+    right_edge = (centers[-1] + dedge_right/2.).reshape(1)
+    edges = np.concatenate((left_edge, midpoints, right_edge))
+    # cleanup nans....
+    w = np.where(~np.isnan(edges))
+    edges = edges[w]
+    if positive:
+        newedges = list()
+        mxedge = 0
+        for edge in edges:
+            if edge > mxedge:
+                newedges.append(edge)
+                mxedge = edge
+        edges = np.array(newedges)
+    return edges
