+++ SciStreams/tests/test_XS_functions.py
@@ -1,6 +1,7 @@
 # test the XSAnalysis Streams, make sure they're working properly
-from SciStreams.analyses.XSAnalysis.Streams import circavg
-from SciStreams.analyses.XSAnalysis.tools import xystitch_accumulate
+from SciStreams.processing.circavg import circavg
+# from SciStreams.processing.stitching import xystitch_accumulate
+# TODO : test all functions
 
 import numpy as np
 
@@ -31,17 +32,3 @@ def test_circavg():
     assert 'sqxerr' in res.kwargs
     assert 'sqy' in res.kwargs
     assert 'sqyerr' in res.kwargs
-
-
-def test_xystitch_accumulate():
-    # mostly make sure it runs with no errors
-    img = np.zeros((100, 100), dtype=int)
-    mask = np.ones((100, 100), dtype=float)
-    origin = (40., 50)
-    stitchback = True
-
-    prevstate = img, mask, origin, stitchback
-    newstate = img, mask, origin, stitchback
-    res = xystitch_accumulate(prevstate, newstate)
-    assert res[0].shape == img.shape
-    assert res[1].shape == mask.shape
