+++ SciStreams/utils/utils.py
